Move format handling into own class
