+++ userspace/digwatch/CMakeLists.txt
@@ -10,7 +10,7 @@ include_directories(${PROJECT_SOURCE_DIR}/../sysdig/userspace/libscap)
 include_directories(${PROJECT_SOURCE_DIR}/../sysdig/userspace/libsinsp)
 include_directories("${PROJECT_BINARY_DIR}/userspace/digwatch")
 
-add_executable(digwatch rules.cpp digwatch.cpp)
+add_executable(digwatch formats.cpp rules.cpp digwatch.cpp)
 
 target_link_libraries(digwatch sinsp)
