+++ userspace/digwatch/rules.cpp
@@ -6,57 +6,15 @@ extern "C" {
 #include "lauxlib.h"
 }
 
-std::map<uint32_t, sinsp_evt_formatter*> g_format_map;
-sinsp* g_inspector;
-
-const static struct luaL_reg ll_digwatch [] =
-{
-	{"set_formatter", &digwatch_rules::set_formatter},
-	{NULL,NULL}
-};
 
 digwatch_rules::digwatch_rules(sinsp* inspector, lua_State *ls, string lua_main_filename, string lua_dir)
 {
-	g_inspector = inspector;
-
 	m_ls = ls;
 
 	m_lua_parser = new lua_parser(inspector, m_ls);
 
-	luaL_openlib(m_ls, "digwatch", ll_digwatch, 0);
-
 	add_lua_path(lua_dir);
 	load_compiler(lua_main_filename);
-}
-
-int digwatch_rules::set_formatter (lua_State *ls) {
-	uint32_t index = luaL_checkinteger(ls, 1);
-	string format = luaL_checkstring(ls, 2);
-
-	try
-	{
-		if(format == "" || format == "default")
-		{
-			g_format_map[index] = new sinsp_evt_formatter(g_inspector, DEFAULT_OUTPUT_STR);
-		}
-		else
-		{
-			g_format_map[index] = new sinsp_evt_formatter(g_inspector, format);
-		}
-	}
-	catch(sinsp_exception& e)
-	{
-		string err = "invalid output format " + format;
-		fprintf(stderr, "%s\n", err.c_str());
-		throw sinsp_exception("set_formatter error");
-	}
-
-	return 0;
-}
-
-sinsp_evt_formatter* digwatch_rules::lookup_formatter(uint32_t index)
-{
-	return g_format_map[index];
 }
 
 void digwatch_rules::add_lua_path(string path)
